+++ ios/Classes/NativeIdPlugin.swift
@@ -1,17 +1,20 @@
 import Flutter
 import UIKit
 
-public class SwiftNativeIdPlugin: NSObject, FlutterPlugin {
+public class NativeIdPlugin: NSObject, FlutterPlugin {
   public static func register(with registrar: FlutterPluginRegistrar) {
     let channel = FlutterMethodChannel(name: "native_id", binaryMessenger: registrar.messenger())
-    let instance = SwiftDeviceInformationPlugin()
+    let instance = NativeIdPlugin()
     registrar.addMethodCallDelegate(instance, channel: channel)
   }
 
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
-    if(call.method.elementsEqual("getId")){
-      let id = UIDevice.current.identifierForVendor?.uuidString
-        result(id)
+    switch call.method {
+    case "getId":
+      let identifierForVendor = UIDevice.current.identifierForVendor?.uuidString
+      result(identifierForVendor)
+    default:
+      result(FlutterMethodNotImplemented)
     }
   }
 }