+++ test/native_id_method_channel_test.dart
@@ -3,19 +3,22 @@
 // import 'package:native_id/native_id_method_channel.dart';
 
 // void main() {
+//   TestWidgetsFlutterBinding.ensureInitialized();
+
 //   MethodChannelNativeId platform = MethodChannelNativeId();
 //   const MethodChannel channel = MethodChannel('native_id');
 
-//   TestWidgetsFlutterBinding.ensureInitialized();
-
 //   setUp(() {
-//     channel.setMockMethodCallHandler((MethodCall methodCall) async {
-//       return '42';
-//     });
+//     TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger.setMockMethodCallHandler(
+//       channel,
+//       (MethodCall methodCall) async {
+//         return '42';
+//       },
+//     );
 //   });
 
 //   tearDown(() {
-//     channel.setMockMethodCallHandler(null);
+//     TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger.setMockMethodCallHandler(channel, null);
 //   });
 
 //   test('getPlatformVersion', () async {